+++ noodles-csi/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Changed
 
-  * csi/index: Change `min_shift` to a `u8`.
+  * csi/index: Change `min_shift` and `depth` to a `u8`.
 
 ## 0.5.1 - 2022-03-02